FE-1244 - exports app layout to the component
